Adiciona função de handleback

+++ views/pages/areaRestrita/index.js
@@ -1,7 +1,7 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { createMaterialBottomTabNavigator } from "@react-navigation/material-bottom-tabs";
 import React, { useEffect, useState } from "react";
-import { View, Text, Button } from "react-native";
+import { BackHandler, Alert } from "react-native";
 
 import Icon from "react-native-vector-icons/Ionicons";
 import { Profile, Edicao, Cadastro } from "../../bridge.js";
@@ -10,10 +10,32 @@ import styles from "./styles.js";
 export default function AreaRestrita({ navigation }) {
   const Tab = createMaterialBottomTabNavigator();
 
-  async function exitProfile() {
-    navigation.navigate("Login");
-    await AsyncStorage.clear();
-  }
+  useEffect(() => {
+    const backAction = () => {
+      Alert.alert("", "Deseja fazer log-out da sua conta?", [
+        {
+          text: "Não",
+          onPress: () => null,
+          style: "não",
+        },
+        {
+          text: "Sim",
+          onPress: () => {
+            AsyncStorage.clear();
+            navigation.navigate("Home");
+          },
+        },
+      ]);
+      return true;
+    };
+
+    const backHandler = BackHandler.addEventListener(
+      "hardwareBackPress",
+      backAction
+    );
+
+    return () => backHandler.remove();
+  }, []);
 
   return (
     <Tab.Navigator
@@ -40,7 +62,9 @@ export default function AreaRestrita({ navigation }) {
         name="Cadastrar"
         component={Cadastro}
         options={{
-          tabBarIcon: () => <Icon name="add-circle" size={20} color="#b2772c" />,
+          tabBarIcon: () => (
+            <Icon name="add-circle" size={20} color="#b2772c" />
+          ),
         }}
       />
     </Tab.Navigator>
